+++ ginisdk/src/androidTest/java/net/gini/android/MediaTypesTest.java
@@ -1,11 +1,16 @@
 package net.gini.android;
 
+import android.support.test.filters.SmallTest;
 import android.test.AndroidTestCase;
 
-
+@SmallTest
 public class MediaTypesTest extends AndroidTestCase {
     public void testGINI_JSON_V1() {
         assertEquals("application/vnd.gini.v1+json", MediaTypes.GINI_JSON_V1);
+    }
+
+    public void testGINI_JSON_V2() {
+        assertEquals("application/vnd.gini.v2+json", MediaTypes.GINI_JSON_V2);
     }
 
     public void testGINI_JSON_INCUBATOR() {